move CoverImageComponent to Developers namespace
