+++ app/components/developers/cover_image_component.html.erb
