+++ config/locales/pt-BR.yml
@@ -111,8 +111,6 @@ pt-BR:
       flag: Sinalizar essa conversa
       title: Conversa
       with: com
-  cover_image_component:
-    alt: Imagem de capa do desenvolvedor
   date:
     formats:
       month_year: "%B / %Y"
@@ -121,6 +119,8 @@ pt-BR:
       in_future_html: Disponível em <time datetime="%{date}">%{date_in_words}</time>
       now_html: Disponível agora
       unspecified_html: Atualmente indisponível
+    cover_image_component:
+      alt: Imagem de capa do desenvolvedor
     create:
       created: Seu perfil foi adicionado!
     details:
